update filters to only show when pueblo view. update label to find pueblo. touch up profile.
