+++ src/app/profile/components/ThingsToDoPueblos.tsx
@@ -1,0 +1,7 @@
+import NotFoundOverlay from "@/components/NotFoundOverlay";
+
+const ThingsToDoPueblos = () => {
+  return <NotFoundOverlay title="Coming soon!" showButton={true} />;
+};
+
+export default ThingsToDoPueblos;
