+++ src/components/filterNavBar.module.css
@@ -1,0 +1,3 @@
+.drawerButton {
+  margin: 1rem 0;
+}
