--- src/components/navbar.module.css
@@ -1,68 +1,0 @@
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-  flex-wrap: wrap;
-}
-
-.titleContainer {
-  display: flex;
-  padding: 1rem;
-}
-
-.leftLinks,
-.rightLinks {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.logo {
-  font-weight: bold;
-  font-size: 1.25rem;
-  text-decoration: none;
-}
-
-.logoImg {
-  height: 1.8rem;
-  width: 1.8rem;
-  align-self: end;
-  margin: 0 0.5rem;
-}
-
-.logoImgSunglasses {
-  height: 1.8rem;
-  width: 1.8rem;
-  align-self: end;
-  margin: 0 0.5rem;
-}
-
-.drawerButton {
-  margin-top: 1rem;
-  margin-right: 1rem;
-}
-
-.link {
-  text-decoration: none;
-  color: #2563eb; /* Tailwind blue-600 */
-  transition: color 0.2s ease;
-}
-
-.link:hover {
-  text-decoration: underline;
-}
-
-.logoutButton {
-  background: none;
-  border: none;
-  color: #dc2626; /* Tailwind red-600 */
-  cursor: pointer;
-  font: inherit;
-  text-decoration: none;
-  transition: color 0.2s ease;
-}
-
-.logoutButton:hover {
-  text-decoration: underline;
-}
